Foreign.CUDA.BLAS.Error

Description

Documentation

data Status

Error codes used by cuBLAS library functions

http://docs.nvidia.com/cuda/cublas/index.html#cublasstatus_t

Constructors

Success

NotInitialized

AllocFailed

InvalidValue

ArchMismatch

MappingError

ExecutionFailed

InternalError

NotSupported

LicenseError

describe :: Status -> String

data CUBLASException

Constructors

ExitCode Status

UserError String

cublasError :: String -> IO a

Raise a CUBLASException in the IO Monad

requireSDK :: Name -> Double -> IO a

A specially formatted error message

resultIfOk :: (Status, a) -> IO a

Return the results of a function on successful execution, otherwise throw an exception with an error string associated with the return code

nothingIfOk :: Status -> IO ()

Throw an exception with an error string associated with an unsuccessful return code, otherwise return unit.

checkStatus :: CInt -> IO ()