Foreign.CUDA.BLAS.Level2

Description

For more information see the cuBLAS Level-2 function reference:

http://docs.nvidia.com/cuda/cublas/index.html#cublas-level-2-function-reference

Documentation

data Operation

Indicates which operation needs to be performed with a dense matrix.

N: no transpose selected

T: transpose operation

C: conjugate transpose

http://docs.nvidia.com/cuda/cublas/index.html#cublasoperation_t

Constructors

N

T

C

data Fill

Indicates which part, upper or lower, of a dense matrix was filled and consequently should be used by the function.

http://docs.nvidia.com/cuda/cublas/index.html#cublasfillmode_t

Constructors

Lower

Upper

data Diagonal

Indicates whether the main diagonal of a dense matrix is unity and consequently should not be be touched or modified by the function.

http://docs.nvidia.com/cuda/cublas/index.html#cublasdiagtype_t

Constructors

NonUnit

Unit

sgbmv :: Handle -> Operation -> Int -> Int -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dgbmv :: Handle -> Operation -> Int -> Int -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

cgbmv :: Handle -> Operation -> Int -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zgbmv :: Handle -> Operation -> Int -> Int -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

sgemv :: Handle -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dgemv :: Handle -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

cgemv :: Handle -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zgemv :: Handle -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

sger :: Handle -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dger :: Handle -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

cgerc :: Handle -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zgerc :: Handle -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

cgeru :: Handle -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zgeru :: Handle -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

ssbmv :: Handle -> Fill -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dsbmv :: Handle -> Fill -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

sspmv :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr Float -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dspmv :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr Double -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

sspr :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> IO ()

dspr :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> IO ()

sspr2 :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> IO ()

dspr2 :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> IO ()

ssymv :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dsymv :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

csymv :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zsymv :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

ssyr :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dsyr :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

csyr :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zsyr :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

ssyr2 :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dsyr2 :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

csyr2 :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zsyr2 :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

stbmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dtbmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

ctbmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

ztbmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

stbsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dtbsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

ctbsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

ztbsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

stpmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Float -> DevicePtr Float -> Int -> IO ()

dtpmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Double -> DevicePtr Double -> Int -> IO ()

ctpmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

ztpmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

stpsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Float -> DevicePtr Float -> Int -> IO ()

dtpsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Double -> DevicePtr Double -> Int -> IO ()

ctpsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

ztpsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

strmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dtrmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

ctrmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

ztrmv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

strsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dtrsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

ctrsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

ztrsv :: Handle -> Fill -> Operation -> Diagonal -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

chemv :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zhemv :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

chbmv :: Handle -> Fill -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zhbmv :: Handle -> Fill -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

chpmv :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zhpmv :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

cher :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zher :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

cher2 :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zher2 :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

chpr :: Handle -> Fill -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> IO ()

zhpr :: Handle -> Fill -> Int -> Ptr Double -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> IO ()

chpr2 :: Handle -> Fill -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> IO ()

zhpr2 :: Handle -> Fill -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> IO ()