Foreign.CUDA.BLAS.Level3

Description

For more information see the cuBLAS Level-3 function reference:

http://docs.nvidia.com/cuda/cublas/index.html#cublas-level-3-function-reference

Documentation

data Operation

Indicates which operation needs to be performed with a dense matrix.

N: no transpose selected

T: transpose operation

C: conjugate transpose

http://docs.nvidia.com/cuda/cublas/index.html#cublasoperation_t

Constructors

N

T

C

data Fill

Indicates which part, upper or lower, of a dense matrix was filled and consequently should be used by the function.

http://docs.nvidia.com/cuda/cublas/index.html#cublasfillmode_t

Constructors

Lower

Upper

data Diagonal

Indicates whether the main diagonal of a dense matrix is unity and consequently should not be be touched or modified by the function.

http://docs.nvidia.com/cuda/cublas/index.html#cublasdiagtype_t

Constructors

NonUnit

Unit

data Side

Indicates whether the dense matrix is on the lift or right side in the matrix equation solved by a particular function.

http://docs.nvidia.com/cuda/cublas/index.html#cublassidemode_t

Constructors

Left

Right

data Type

Used to specify the type of data underlying a void* pointer. For example, it is used in the routine ?gemmEx.

http://docs.nvidia.com/cuda/cublas/index.html#cuda_datatype_t

Constructors

R32f

R64f

R16f

R8i

C32f

C64f

C16f

C8i

R8u

C8u

R32i

C32i

R32u

C32u

data GemmAlgorithm

Used to run gemmEx with a specific, yet completely unspecified, algorithm.

Constructors

GemmDefault

GemmAlgo0

GemmAlgo1

GemmAlgo2

GemmAlgo3

GemmAlgo4

GemmAlgo5

GemmAlgo6

GemmAlgo7

sgemm :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dgemm :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

cgemm :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zgemm :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

sgemmBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Float -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr (DevicePtr Float) -> Int -> Ptr Float -> DevicePtr (DevicePtr Float) -> Int -> Int -> IO ()

dgemmBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Double -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr (DevicePtr Double) -> Int -> Ptr Double -> DevicePtr (DevicePtr Double) -> Int -> Int -> IO ()

cgemmBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> Ptr (Complex Float) -> DevicePtr (DevicePtr (Complex Float)) -> Int -> Int -> IO ()

zgemmBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> Ptr (Complex Double) -> DevicePtr (DevicePtr (Complex Double)) -> Int -> Int -> IO ()

ssymm :: Handle -> Side -> Fill -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dsymm :: Handle -> Side -> Fill -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

csymm :: Handle -> Side -> Fill -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zsymm :: Handle -> Side -> Fill -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

ssyrk :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dsyrk :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

csyrk :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zsyrk :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

ssyr2k :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dsyr2k :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

csyr2k :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zsyr2k :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

ssyrkx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

dsyrkx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

csyrkx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zsyrkx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

strmm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dtrmm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

ctrmm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

ztrmm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

strsm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dtrsm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

ctrsm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

ztrsm :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

strsmBatched :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr Float -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr (DevicePtr Float) -> Int -> Int -> IO ()

dtrsmBatched :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr Double -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr (DevicePtr Double) -> Int -> Int -> IO ()

ctrsmBatched :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> Int -> IO ()

ztrsmBatched :: Handle -> Side -> Fill -> Operation -> Diagonal -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> Int -> IO ()

chemm :: Handle -> Side -> Fill -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zhemm :: Handle -> Side -> Fill -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

cherk :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Int -> IO ()

zherk :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Double -> DevicePtr (Complex Double) -> Int -> Ptr Double -> DevicePtr (Complex Double) -> Int -> IO ()

cher2k :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Int -> IO ()

zher2k :: Handle -> Fill -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr Double -> DevicePtr (Complex Double) -> Int -> IO ()

cherkx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> IO ()

zherkx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> IO ()

sgeam :: Handle -> Operation -> Operation -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> Ptr Float -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

dgeam :: Handle -> Operation -> Operation -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> Ptr Double -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

cgeam :: Handle -> Operation -> Operation -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zgeam :: Handle -> Operation -> Operation -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

sdgmm :: Handle -> Side -> Int -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> Int -> IO ()

ddgmm :: Handle -> Side -> Int -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> Int -> IO ()

cdgmm :: Handle -> Side -> Int -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> IO ()

zdgmm :: Handle -> Side -> Int -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> IO ()

sgetrfBatched :: Handle -> Int -> DevicePtr (DevicePtr Float) -> Int -> Ptr Int32 -> Ptr Int32 -> Int -> IO ()

dgetrfBatched :: Handle -> Int -> DevicePtr (DevicePtr Double) -> Int -> Ptr Int32 -> Ptr Int32 -> Int -> IO ()

cgetrfBatched :: Handle -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> Ptr Int32 -> Ptr Int32 -> Int -> IO ()

zgetrfBatched :: Handle -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> Ptr Int32 -> Ptr Int32 -> Int -> IO ()

sgetriBatched :: Handle -> Int -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr Int32 -> Int -> IO ()

dgetriBatched :: Handle -> Int -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr Int32 -> Int -> IO ()

cgetriBatched :: Handle -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr Int32 -> Int -> IO ()

zgetriBatched :: Handle -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr Int32 -> Int -> IO ()

smatinvBatched :: Handle -> Int -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr Int32 -> Int -> IO ()

dmatinvBatched :: Handle -> Int -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr Int32 -> Int -> IO ()

cmatinvBatched :: Handle -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr Int32 -> Int -> IO ()

zmatinvBatched :: Handle -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr Int32 -> Int -> IO ()

sgeqrfBatched :: Handle -> Int -> Int -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr (DevicePtr Float) -> HostPtr Int32 -> Int -> IO ()

dgeqrfBatched :: Handle -> Int -> Int -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr (DevicePtr Double) -> HostPtr Int32 -> Int -> IO ()

cgeqrfBatched :: Handle -> Int -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr (DevicePtr (Complex Float)) -> HostPtr Int32 -> Int -> IO ()

zgeqrfBatched :: Handle -> Int -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr (DevicePtr (Complex Double)) -> HostPtr Int32 -> Int -> IO ()

sgelsBatched :: Handle -> Operation -> Int -> Int -> Int -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr (DevicePtr Float) -> Int -> HostPtr Int32 -> DevicePtr Int32 -> Int -> IO ()

dgelsBatched :: Handle -> Operation -> Int -> Int -> Int -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr (DevicePtr Double) -> Int -> HostPtr Int32 -> DevicePtr Int32 -> Int -> IO ()

cgelsBatched :: Handle -> Operation -> Int -> Int -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> HostPtr Int32 -> DevicePtr Int32 -> Int -> IO ()

zgelsBatched :: Handle -> Operation -> Int -> Int -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> HostPtr Int32 -> DevicePtr Int32 -> Int -> IO ()

stpttr :: Handle -> Fill -> Int -> DevicePtr Float -> DevicePtr Float -> Int -> IO ()

dtpttr :: Handle -> Fill -> Int -> DevicePtr Double -> DevicePtr Double -> Int -> IO ()

ctpttr :: Handle -> Fill -> Int -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

ztpttr :: Handle -> Fill -> Int -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

strttp :: Handle -> Fill -> Int -> DevicePtr Float -> Int -> DevicePtr Float -> IO ()

dtrttp :: Handle -> Fill -> Int -> DevicePtr Double -> Int -> DevicePtr Double -> IO ()

ctrttp :: Handle -> Fill -> Int -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> IO ()

ztrttp :: Handle -> Fill -> Int -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> IO ()

sgetrsBatched :: Handle -> Operation -> Int -> Int -> DevicePtr (DevicePtr Float) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr Float) -> Int -> HostPtr Int32 -> Int -> IO ()

dgetrsBatched :: Handle -> Operation -> Int -> Int -> DevicePtr (DevicePtr Double) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr Double) -> Int -> HostPtr Int32 -> Int -> IO ()

cgetrsBatched :: Handle -> Operation -> Int -> Int -> DevicePtr (DevicePtr (Complex Float)) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr (Complex Float)) -> Int -> HostPtr Int32 -> Int -> IO ()

zgetrsBatched :: Handle -> Operation -> Int -> Int -> DevicePtr (DevicePtr (Complex Double)) -> Int -> DevicePtr Int32 -> DevicePtr (DevicePtr (Complex Double)) -> Int -> HostPtr Int32 -> Int -> IO ()

hgemm :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Half -> DevicePtr Half -> Int -> DevicePtr Half -> Int -> Ptr Half -> DevicePtr Half -> Int -> IO ()

sgemmEx :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Float -> DevicePtr () -> Type -> Int -> DevicePtr () -> Type -> Int -> Ptr Float -> DevicePtr () -> Type -> Int -> IO ()

cgemm3m :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> DevicePtr (Complex Float) -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> IO ()

zgemm3m :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> DevicePtr (Complex Double) -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> IO ()

sgemmStridedBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Float -> DevicePtr Float -> Int -> Int64 -> DevicePtr Float -> Int -> Int64 -> Ptr Float -> DevicePtr Float -> Int -> Int64 -> Int -> IO ()

dgemmStridedBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Double -> DevicePtr Double -> Int -> Int64 -> DevicePtr Double -> Int -> Int64 -> Ptr Double -> DevicePtr Double -> Int -> Int64 -> Int -> IO ()

cgemmStridedBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Int64 -> DevicePtr (Complex Float) -> Int -> Int64 -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Int64 -> Int -> IO ()

zgemmStridedBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> Int64 -> DevicePtr (Complex Double) -> Int -> Int64 -> Ptr (Complex Double) -> DevicePtr (Complex Double) -> Int -> Int64 -> Int -> IO ()

hgemmStridedBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr Half -> DevicePtr Half -> Int -> Int64 -> DevicePtr Half -> Int -> Int64 -> Ptr Half -> DevicePtr Half -> Int -> Int64 -> Int -> IO ()

cgemm3mStridedBatched :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Int64 -> DevicePtr (Complex Float) -> Int -> Int64 -> Ptr (Complex Float) -> DevicePtr (Complex Float) -> Int -> Int64 -> Int -> IO ()

cgemmEx :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr (Complex Float) -> DevicePtr () -> Type -> Int -> DevicePtr () -> Type -> Int -> Ptr (Complex Float) -> DevicePtr () -> Type -> Int -> IO ()

gemmEx :: Handle -> Operation -> Operation -> Int -> Int -> Int -> Ptr () -> DevicePtr () -> Type -> Int -> DevicePtr () -> Type -> Int -> Ptr () -> DevicePtr () -> Type -> Int -> Type -> GemmAlgorithm -> IO ()

csyrkEx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr () -> Type -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Type -> Int -> IO ()

csyrk3mEx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr () -> Type -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Type -> Int -> IO ()

cherkEx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr () -> Type -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Type -> Int -> IO ()

cherk3mEx :: Handle -> Fill -> Operation -> Int -> Int -> Ptr Float -> DevicePtr () -> Type -> Int -> Ptr Float -> DevicePtr (Complex Float) -> Type -> Int -> IO ()

nrm2Ex :: Handle -> Int -> DevicePtr () -> Type -> Int -> Ptr () -> Type -> Type -> IO ()

axpyEx :: Handle -> Int -> Ptr () -> Type -> DevicePtr () -> Type -> Int -> DevicePtr () -> Type -> Int -> Type -> IO ()

dotEx :: Handle -> Int -> DevicePtr () -> Type -> Int -> DevicePtr () -> Type -> Int -> Ptr () -> Type -> Type -> IO ()

dotcEx :: Handle -> Int -> DevicePtr () -> Type -> Int -> DevicePtr () -> Type -> Int -> Ptr () -> Type -> Type -> IO ()

scalEx :: Handle -> Int -> Ptr () -> Type -> DevicePtr () -> Type -> Int -> Type -> IO ()