Foreign.CUDA.Cublas.Error

Documentation

describe :: Status -> String

data CUBLASException

Constructors

ExitCode Status

UserError String

cublasError :: String -> IO a

Raise a CUBLASException in the IO Monad

resultIfOk :: (Status, a) -> IO a

Return the results of a function on successful execution, otherwise throw an exception with an error string associated with the return code